The delete action now works for Post objects although I'm not convinced it's being done correctly.

+++ BlogEngine.UnitTests/TestPostRepository.cs
@@ -158,7 +158,7 @@ namespace BlogEngine.UnitTests
             return;
         }
 
-        public void DeletePostByID(int id)
+        public void DeletePostByID(long id)
         {
             return;
         }
